Calendar.tsx is at 100% coverage
